# openapi.yaml
openapi: 3.1.0
info:
  title: Mahshar V2 Agent API
  version: 2.1.0
  description: |
    Public machine interface for discovering and purchasing proxied APIs with
    Circle Gateway x402 v2 on Arc Mainnet (chain 5042, CAIP-2 eip155:5042).
    Seller upstream URLs and credentials are never part of this contract.
servers:
  - url: https://mahshar.xyz
...
paths:
...
  /api/proxy:
    post:
      tags: [Paid execution]
      operationId: executeEnvelopeListing
      summary: Probe or execute any listing through the envelope proxy
      description: |
        The outer method is POST. `method` is the configured upstream method.
        `path` is an optional relative suffix. First omit Payment-Signature. A
        402 has body `{}` and a base64 x402 v2 challenge in PAYMENT-REQUIRED.
        Retry the identical URL and JSON body with Payment-Signature.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnvelopeRequest' }
      responses:
        '200': { $ref: '#/components/responses/PaidResult' }
        '400': { $ref: '#/components/responses/Error' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '409': { $ref: '#/components/responses/DeliveryConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/PaidResult' }
        '503': { $ref: '#/components/responses/Error' }
  /api/proxy/{api_id}:
    parameters:
      - name: api_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Paid execution]
      operationId: executeGetListing
      summary: Probe or execute a GET listing
      responses:
        '200': { $ref: '#/components/responses/PaidResult' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '405': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/DeliveryConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Paid execution]
      operationId: executePostListing
      summary: Probe or execute a POST listing
      description: JSON is parsed and re-serialized before being sent upstream.
      requestBody:
        required: false
        content:
          application/json:
            schema: {}
      responses:
        '200': { $ref: '#/components/responses/PaidResult' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '405': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/DeliveryConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/calls/last-response:
    get:
      tags: [Purchase access]
      operationId: retrievePurchasedResponse
      summary: Retrieve the private response for the exact authorized purchase
      description: |
        Send `x-mahshar-purchase-access` with the purchase-scoped capability.
        A legacy purchase may exchange one one-use `x-mahshar-authorization`
        wallet proof for a capability. `buyer_wallet` alone is never ownership.
      parameters:
        - { name: api_id, in: query, required: true, schema: { type: string } }
        - { name: buyer_wallet, in: query, required: true, schema: { type: string } }
        - { name: x-mahshar-purchase-access, in: header, required: false, schema: { type: string } }
        - { name: x-mahshar-authorization, in: header, required: false, schema: { type: string } }
      responses:
        '200':
          description: Stored response and optionally a newly issued capability
          content:
            application/json:
              schema:
                type: object
                properties:
                  response_body: {}
                  purchase_access_token: { type: string }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  schemas:
    EnvelopeRequest:
      type: object
      additionalProperties: false
      required: [api_id, buyer_wallet]
      properties:
        api_id: { type: string, format: uuid }
        buyer_wallet:
          type: string
          description: Probe hint only; the verified settlement payer is authoritative.
        method: { type: string, enum: [GET, POST, PUT, DELETE] }
        path: { type: string, description: Relative suffix when discovery declares dynamic path support. }
        body: { description: JSON for POST, PUT, or DELETE; GET never forwards a body. }
...
    PaidResult:
      type: object
      required: [payment, delivery_state, attemptId, purchase_access_token]
      properties:
        response: { description: Upstream JSON value or text string. }
        latency_ms: { type: integer }
        payment: { const: ACCOUNTING_COMPLETE }
        delivery_state: { enum: [SUCCEEDED, FAILED_RETRYABLE, FAILED_FINAL, UNKNOWN] }
        retryable: { type: boolean }
        attemptId: { type: string, format: uuid }
        purchase_access_token: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        attemptId: { type: string }
        delivery_state: { type: string }
        retryable: { type: boolean }
...
  responses:
    PaymentRequired:
      description: Payment required; body is exactly `{}`.
      headers:
        PAYMENT-REQUIRED:
          required: true
          description: Base64-encoded x402 v2 PaymentRequired JSON.
          schema: { type: string }
      content:
        application/json:
          schema: { type: object, maxProperties: 0 }
    PaidResult:
      description: |
        Settlement succeeded. PAYMENT-RESPONSE is a standard base64 x402 v2
        SettleResponse even when upstream delivery is non-2xx. An upstream body
        whose final serialized JSON wrapper exceeds 4,000,000 bytes produces 502 and FAILED_FINAL.
      headers:
        PAYMENT-RESPONSE:
          required: true
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PaidResult' }
    Error:
      description: Machine-readable error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    DeliveryConflict:
      description: Payment accounted; delivery pending, final, or unknown. Do not authorize a second payment.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/Error' }
              - type: object
                properties:
                  error: { const: rate_limited }
                  retry_after_seconds: { type: integer }